Refactor to Core
